move all changes back into the original files
