+++ git_sim/git_sim_base_command.py
@@ -12,7 +12,6 @@ from git_sim.settings import Settings
 
 class GitSimBaseCommand(m.MovingCameraScene):
     def __init__(self):
-
         super().__init__()
         self.init_repo()
 
@@ -136,7 +135,6 @@ class GitSimBaseCommand(m.MovingCameraScene):
 
     def show_outro(self):
         if Settings.animate and Settings.show_outro:
-
             self.play(m.Restore(self.camera.frame))
 
             self.play(self.logo.animate.scale(4).set_x(0).set_y(0))
@@ -364,7 +362,6 @@ class GitSimBaseCommand(m.MovingCameraScene):
             return
 
         for tag in self.repo.tags:
-
             try:
                 if commit.hexsha == tag.commit.hexsha:
                     tagText = m.Text(
@@ -741,7 +738,6 @@ class GitSimBaseCommand(m.MovingCameraScene):
         firstColumnArrowMap={},
         secondColumnArrowMap={},
     ):
-
         for x in self.repo.index.diff(None):
             if "git-sim_media" not in x.a_path:
                 secondColumnFileNames.add(x.a_path)
