+++ git_sim/git_sim_branch.py
@@ -1,16 +1,18 @@
-from argparse import Namespace
-
 import manim as m
+import typer
 
+from git_sim.animations import handle_animations
 from git_sim.git_sim_base_command import GitSimBaseCommand
+from git_sim.settings import Settings
 
 
-class GitSimBranch(GitSimBaseCommand):
-    def __init__(self, args: Namespace):
-        super().__init__(args=args)
+class Branch(GitSimBaseCommand):
+    def __init__(self, name: str):
+        super().__init__()
+        self.name = name
 
     def construct(self):
-        print("Simulating: git " + self.args.subcommand + " " + self.args.name)
+        print(f"{Settings.INFO_STRING} branch {self.name}")
 
         self.show_intro()
         self.get_commits()
@@ -19,7 +21,7 @@ class GitSimBranch(GitSimBaseCommand):
         self.scale_frame()
 
         branchText = m.Text(
-            self.args.name,
+            self.name,
             font="Monospace",
             font_size=20,
             color=self.fontColor,
@@ -37,13 +39,23 @@ class GitSimBranch(GitSimBaseCommand):
 
         fullbranch = m.VGroup(branchRec, branchText)
 
-        if self.args.animate:
-            self.play(m.Create(fullbranch), run_time=1 / self.args.speed)
+        if Settings.animate:
+            self.play(m.Create(fullbranch), run_time=1 / Settings.speed)
         else:
             self.add(fullbranch)
 
         self.toFadeOut.add(branchRec, branchText)
-        self.drawnRefs[self.args.name] = fullbranch
+        self.drawnRefs[self.name] = fullbranch
 
         self.fadeout()
         self.show_outro()
+
+
+def branch(
+    name: str = typer.Argument(
+        ...,
+        help="The name of the new branch",
+    )
+):
+    scene = Branch(name=name)
+    handle_animations(scene=scene)
