+++ git_sim/git_sim_cherrypick.py
@@ -1,28 +1,32 @@
 import sys
-from argparse import Namespace
 
 import git
 import manim as m
+import typer
 
+from git_sim.animations import handle_animations
 from git_sim.git_sim_base_command import GitSimBaseCommand
+from git_sim.settings import Settings
 
 
-class GitSimCherryPick(GitSimBaseCommand):
-    def __init__(self, args: Namespace):
-        super().__init__(args=args)
+class CherryPick(GitSimBaseCommand):
+    def __init__(self, commit: str, edit: str):
+        super().__init__()
+        self.commit = commit
+        self.edit = edit
 
         try:
-            git.repo.fun.rev_parse(self.repo, self.args.commit[0])
+            git.repo.fun.rev_parse(self.repo, self.commit)
         except git.exc.BadName:
             print(
                 "git-sim error: '"
-                + self.args.commit[0]
+                + self.commit
                 + "' is not a valid Git ref or identifier."
             )
             sys.exit(1)
 
-        if self.args.commit[0] in [branch.name for branch in self.repo.heads]:
-            self.selected_branches.append(self.args.commit[0])
+        if self.commit in [branch.name for branch in self.repo.heads]:
+            self.selected_branches.append(self.commit)
 
         try:
             self.selected_branches.append(self.repo.active_branch.name)
@@ -30,20 +34,17 @@ class GitSimCherryPick(GitSimBaseCommand):
             pass
 
     def construct(self):
-        print(
-            "Simulating: git "
-            + self.args.subcommand
-            + " "
-            + self.args.commit[0]
-            + ((' -e "' + self.args.edit + '"') if self.args.edit else "")
-        )
+        substring = ""
+        if self.edit:
+            substring = f' -e "{self.edit}"'
+        print(f"{Settings.INFO_STRING} cherrypick {self.commit}{substring}")
 
         if self.repo.active_branch.name in self.repo.git.branch(
-            "--contains", self.args.commit[0]
+            "--contains", self.commit
         ):
             print(
                 "git-sim error: Commit '"
-                + self.args.commit[0]
+                + self.commit
                 + "' is already included in the history of active branch '"
                 + self.repo.active_branch.name
                 + "'."
@@ -54,12 +55,12 @@ class GitSimCherryPick(GitSimBaseCommand):
         self.get_commits()
         self.parse_commits(self.commits[0])
         self.orig_commits = self.commits
-        self.get_commits(start=self.args.commit[0])
+        self.get_commits(start=self.commit)
         self.parse_commits(self.commits[0], shift=4 * m.DOWN)
         self.center_frame_on_commit(self.orig_commits[0])
         self.setup_and_draw_parent(
             self.orig_commits[0],
-            self.args.edit if self.args.edit else self.commits[0].message,
+            self.edit if self.edit else self.commits[0].message,
         )
         self.draw_arrow_between_commits(self.commits[0].hexsha, "abcdef")
         self.recenter_frame()
@@ -67,3 +68,17 @@ class GitSimCherryPick(GitSimBaseCommand):
         self.reset_head_branch("abcdef")
         self.fadeout()
         self.show_outro()
+
+
+def cherrypick(
+    commit: str = typer.Argument(
+        ...,
+        help="The ref (branch/tag), or commit ID to simulate cherry-pick onto active branch",
+    ),
+    edit: str = typer.Option(
+        default=None,
+        help="Specify a new commit message for the cherry-picked commit",
+    ),
+):
+    scene = CherryPick(commit=commit, edit=edit)
+    handle_animations(scene=scene)
