+++ git_sim/git_sim_log.py
@@ -1,21 +1,22 @@
-from argparse import Namespace
+import typer
 
+from git_sim.animations import handle_animations
 from git_sim.git_sim_base_command import GitSimBaseCommand
+from git_sim.settings import Settings
 
 
-class GitSimLog(GitSimBaseCommand):
-    def __init__(self, args: Namespace):
-        super().__init__(args=args)
-        self.numCommits = self.args.commits + 1
-        self.defaultNumCommits = self.args.commits + 1
+class Log(GitSimBaseCommand):
+    def __init__(self, commits: int):
+        super().__init__()
+        self.numCommits = commits + 1
+        self.defaultNumCommits = commits + 1
         try:
             self.selected_branches.append(self.repo.active_branch.name)
         except TypeError:
             pass
 
     def construct(self):
-        print("Simulating: git " + self.args.subcommand)
-
+        print(Settings.INFO_STRING + type(self).__name__)
         self.show_intro()
         self.get_commits()
         self.parse_commits(self.commits[0])
@@ -23,3 +24,15 @@ class GitSimLog(GitSimBaseCommand):
         self.scale_frame()
         self.fadeout()
         self.show_outro()
+
+
+def log(
+    commits: int = typer.Option(
+        default=Settings.commits,
+        help="The number of commits to display in the simulated log output",
+        min=1,
+        max=12,
+    ),
+):
+    scene = Log(commits=commits)
+    handle_animations(scene=scene)
