+++ git_sim/git_sim_merge.py
@@ -4,27 +4,32 @@ from argparse import Namespace
 import git
 import manim as m
 import numpy
+import typer
 
+from git_sim.animations import handle_animations
 from git_sim.git_sim_base_command import GitSimBaseCommand
+from git_sim.settings import Settings
 
 
-class GitSimMerge(GitSimBaseCommand):
-    def __init__(self, args: Namespace):
-        super().__init__(args=args)
+class Merge(GitSimBaseCommand):
+    def __init__(self, branch: str, no_ff: bool):
+        super().__init__()
+        self.branch = branch
+        self.no_ff = no_ff
 
         try:
-            git.repo.fun.rev_parse(self.repo, self.args.branch[0])
+            git.repo.fun.rev_parse(self.repo, self.branch)
         except git.exc.BadName:
             print(
                 "git-sim error: '"
-                + self.args.branch[0]
+                + self.branch
                 + "' is not a valid Git ref or identifier."
             )
             sys.exit(1)
 
         self.ff = False
-        if self.args.branch[0] in [branch.name for branch in self.repo.heads]:
-            self.selected_branches.append(self.args.branch[0])
+        if self.branch in [branch.name for branch in self.repo.heads]:
+            self.selected_branches.append(self.branch)
 
         try:
             self.selected_branches.append(self.repo.active_branch.name)
@@ -32,14 +37,14 @@ class GitSimMerge(GitSimBaseCommand):
             pass
 
     def construct(self):
-        print("Simulating: git " + self.args.subcommand + " " + self.args.branch[0])
+        print(Settings.INFO_STRING + "merge " + self.branch)
 
         if self.repo.active_branch.name in self.repo.git.branch(
-            "--contains", self.args.branch[0]
+            "--contains", self.branch
         ):
             print(
                 "git-sim error: Branch '"
-                + self.args.branch[0]
+                + self.branch
                 + "' is already included in the history of active branch '"
                 + self.repo.active_branch.name
                 + "'."
@@ -49,27 +54,27 @@ class GitSimMerge(GitSimBaseCommand):
         self.show_intro()
         self.get_commits()
         self.orig_commits = self.commits
-        self.get_commits(start=self.args.branch[0])
+        self.get_commits(start=self.branch)
 
         # Use forward slash to determine if supplied branch arg is local or remote tracking branch
-        if not self.is_remote_tracking_branch(self.args.branch[0]):
-            if self.args.branch[0] in self.repo.git.branch(
+        if not self.is_remote_tracking_branch(self.branch):
+            if self.branch in self.repo.git.branch(
                 "--contains", self.orig_commits[0].hexsha
             ):
                 self.ff = True
         else:
-            if self.args.branch[0] in self.repo.git.branch(
+            if self.branch in self.repo.git.branch(
                 "-r", "--contains", self.orig_commits[0].hexsha
             ):
                 self.ff = True
 
         if self.ff:
-            self.get_commits(start=self.args.branch[0])
+            self.get_commits(start=self.branch)
             self.parse_commits(self.commits[0])
             reset_head_to = self.commits[0].hexsha
             shift = numpy.array([0.0, 0.6, 0.0])
 
-            if self.args.no_ff:
+            if self.no_ff:
                 self.center_frame_on_commit(self.commits[0])
                 commitId = self.setup_and_draw_parent(self.commits[0], "Merge commit")
                 reset_head_to = "abcdef"
@@ -80,9 +85,7 @@ class GitSimMerge(GitSimBaseCommand):
             if "HEAD" in self.drawnRefs:
                 self.reset_head_branch(reset_head_to, shift=shift)
             else:
-                self.draw_ref(
-                    self.commits[0], commitId if self.args.no_ff else self.topref
-                )
+                self.draw_ref(self.commits[0], commitId if self.no_ff else self.topref)
                 self.draw_ref(
                     self.commits[0],
                     self.drawnRefs["HEAD"],
@@ -94,7 +97,7 @@ class GitSimMerge(GitSimBaseCommand):
             self.get_commits()
             self.parse_commits(self.commits[0])
             self.i = 0
-            self.get_commits(start=self.args.branch[0])
+            self.get_commits(start=self.branch)
             self.parse_commits(self.commits[0], shift=4 * m.DOWN)
             self.center_frame_on_commit(self.orig_commits[0])
             self.setup_and_draw_parent(
@@ -112,3 +115,18 @@ class GitSimMerge(GitSimBaseCommand):
 
         self.fadeout()
         self.show_outro()
+
+
+def merge(
+    branch: str = typer.Argument(
+        ...,
+        help="The name of the branch to merge into the active checked-out branch",
+    ),
+    no_ff: bool = typer.Option(
+        False,
+        "--no-ff",
+        help="Simulate creation of a merge commit in all cases, even when the merge could instead be resolved as a fast-forward",
+    ),
+):
+    scene = Merge(branch=branch, no_ff=no_ff)
+    handle_animations(scene=scene)
