+++ git_sim/git_sim_restore.py
@@ -1,36 +1,33 @@
 import sys
-from argparse import Namespace
 
 import manim as m
+import typer
 
+from git_sim.animations import handle_animations
 from git_sim.git_sim_base_command import GitSimBaseCommand
+from git_sim.settings import Settings
 
 
-class GitSimRestore(GitSimBaseCommand):
-    def __init__(self, args: Namespace):
-        super().__init__(args=args)
+class Restore(GitSimBaseCommand):
+    def __init__(self, files: list[str]):
+        super().__init__()
         self.hide_first_tag = True
+        self.files = files
 
         try:
             self.selected_branches.append(self.repo.active_branch.name)
         except TypeError:
             pass
 
-        for name in self.args.name:
-            if name not in [x.a_path for x in self.repo.index.diff(None)] + [
+        for file in self.files:
+            if file not in [x.a_path for x in self.repo.index.diff(None)] + [
                 y.a_path for y in self.repo.index.diff("HEAD")
             ]:
-                print(
-                    "git-sim error: No modified or staged file with name: '"
-                    + name
-                    + "'"
-                )
+                print(f"git-sim error: No modified or staged file with name: '{file}'")
                 sys.exit()
 
     def construct(self):
-        print(
-            "Simulating: git " + self.args.subcommand + " " + " ".join(self.args.name)
-        )
+        print(Settings.INFO_STRING + "restore " + " ".join(self.files))
 
         self.show_intro()
         self.get_commits()
@@ -50,12 +47,11 @@ class GitSimRestore(GitSimBaseCommand):
         firstColumnArrowMap,
         secondColumnArrowMap,
     ):
-
         for x in self.repo.index.diff(None):
             if "git-sim_media" not in x.a_path:
                 secondColumnFileNames.add(x.a_path)
-                for name in self.args.name:
-                    if name == x.a_path:
+                for file in self.files:
+                    if file == x.a_path:
                         thirdColumnFileNames.add(x.a_path)
                         secondColumnArrowMap[x.a_path] = m.Arrow(
                             stroke_width=3, color=self.fontColor
@@ -64,9 +60,19 @@ class GitSimRestore(GitSimBaseCommand):
         for y in self.repo.index.diff("HEAD"):
             if "git-sim_media" not in y.a_path:
                 firstColumnFileNames.add(y.a_path)
-                for name in self.args.name:
-                    if name == y.a_path:
+                for file in self.files:
+                    if file == y.a_path:
                         secondColumnFileNames.add(y.a_path)
                         firstColumnArrowMap[y.a_path] = m.Arrow(
                             stroke_width=3, color=self.fontColor
                         )
+
+
+def restore(
+    files: list[str] = typer.Argument(
+        default=None,
+        help="The names of one or more files to restore",
+    )
+):
+    scene = Restore(files=files)
+    handle_animations(scene=scene)
