+++ git_sim/git_sim_stash.py
@@ -1,43 +1,37 @@
 import sys
-from argparse import Namespace
 
-import git
 import manim as m
-import numpy
+import typer
 
+from git_sim.animations import handle_animations
 from git_sim.git_sim_base_command import GitSimBaseCommand
 
 
-class GitSimStash(GitSimBaseCommand):
-    def __init__(self, args: Namespace):
-        super().__init__(args=args)
+class Stash(GitSimBaseCommand):
+    def __init__(self, files: list[str]):
+        super().__init__()
         self.hide_first_tag = True
+        self.files = files
 
         try:
             self.selected_branches.append(self.repo.active_branch.name)
         except TypeError:
             pass
 
-        for name in self.args.name:
-            if name not in [x.a_path for x in self.repo.index.diff(None)] + [
+        for file in self.files:
+            if file not in [x.a_path for x in self.repo.index.diff(None)] + [
                 y.a_path for y in self.repo.index.diff("HEAD")
             ]:
-                print(
-                    "git-sim error: No modified or staged file with name: '"
-                    + name
-                    + "'"
-                )
+                print(f"git-sim error: No modified or staged file with name: '{file}'")
                 sys.exit()
 
-        if not self.args.name:
-            self.args.name = [x.a_path for x in self.repo.index.diff(None)] + [
+        if not self.files:
+            self.files = [x.a_path for x in self.repo.index.diff(None)] + [
                 y.a_path for y in self.repo.index.diff("HEAD")
             ]
 
     def construct(self):
-        print(
-            "Simulating: git " + self.args.subcommand + " " + " ".join(self.args.name)
-        )
+        print("Simulating: git stash " + " ".join(self.files))
 
         self.show_intro()
         self.get_commits()
@@ -61,11 +55,10 @@ class GitSimStash(GitSimBaseCommand):
         firstColumnArrowMap,
         secondColumnArrowMap,
     ):
-
         for x in self.repo.index.diff(None):
             firstColumnFileNames.add(x.a_path)
-            for name in self.args.name:
-                if name == x.a_path:
+            for file in self.files:
+                if file == x.a_path:
                     thirdColumnFileNames.add(x.a_path)
                     firstColumnArrowMap[x.a_path] = m.Arrow(
                         stroke_width=3, color=self.fontColor
@@ -73,9 +66,19 @@ class GitSimStash(GitSimBaseCommand):
 
         for y in self.repo.index.diff("HEAD"):
             secondColumnFileNames.add(y.a_path)
-            for name in self.args.name:
-                if name == y.a_path:
+            for file in self.files:
+                if file == y.a_path:
                     thirdColumnFileNames.add(y.a_path)
                     secondColumnArrowMap[y.a_path] = m.Arrow(
                         stroke_width=3, color=self.fontColor
                     )
+
+
+def stash(
+    files: list[str] = typer.Argument(
+        default=None,
+        help="The name of the file to stash changes for",
+    )
+):
+    scene = Stash(files=files)
+    handle_animations(scene=scene)
