+++ git_sim/git_sim_status.py
@@ -1,22 +1,17 @@
-from argparse import Namespace
-
+from git_sim.animations import handle_animations
 from git_sim.git_sim_base_command import GitSimBaseCommand
+from git_sim.settings import Settings
 
 
-class GitSimStatus(GitSimBaseCommand):
-    def __init__(self, args: Namespace):
-        super().__init__(args=args)
-        self.hide_first_tag = True
-        self.allow_no_commits = True
-
+class Status(GitSimBaseCommand):
+    def __init__(self):
+        super().__init__()
         try:
             self.selected_branches.append(self.repo.active_branch.name)
         except TypeError:
             pass
 
     def construct(self):
-        print("Simulating: git " + self.args.subcommand)
-
         self.show_intro()
         self.get_commits()
         self.parse_commits(self.commits[0])
@@ -26,3 +21,11 @@ class GitSimStatus(GitSimBaseCommand):
         self.setup_and_draw_zones()
         self.fadeout()
         self.show_outro()
+
+
+def status():
+    Settings.hide_first_tag = True
+    Settings.allow_no_commits = True
+
+    scene = Status()
+    handle_animations(scene=scene)
